+++ common/src/main/java/net/sashiro/compressedblocks/block/BlockList.java
@@ -234,27 +234,27 @@ public class BlockList {
     public static final CBBlock[] PURPLE_CONCRETE = createBlocks(null);
     public static final CBBlock[] MAGENTA_CONCRETE = createBlocks(null);
     public static final CBBlock[] PINK_CONCRETE = createBlocks(null);
-    public static final CBBlock[] GLASS = createBlocks(0.125F);
-    public static final CBBlock[] TINTED_GLASS = createBlocks(0.25F);
-    public static final CBBlock[] WHITE_STAINED_GLASS = createBlocks(0.125F);
-    public static final CBBlock[] LIGHT_GRAY_STAINED_GLASS = createBlocks(0.125F);
-    public static final CBBlock[] GRAY_STAINED_GLASS = createBlocks(0.125F);
-    public static final CBBlock[] BLACK_STAINED_GLASS = createBlocks(0.125F);
-    public static final CBBlock[] BROWN_STAINED_GLASS = createBlocks(0.125F);
-    public static final CBBlock[] RED_STAINED_GLASS = createBlocks(0.125F);
-    public static final CBBlock[] ORANGE_STAINED_GLASS = createBlocks(0.125F);
-    public static final CBBlock[] YELLOW_STAINED_GLASS = createBlocks(0.125F);
-    public static final CBBlock[] LIME_STAINED_GLASS = createBlocks(0.125F);
-    public static final CBBlock[] GREEN_STAINED_GLASS = createBlocks(0.125F);
-    public static final CBBlock[] CYAN_STAINED_GLASS = createBlocks(0.125F);
-    public static final CBBlock[] LIGHT_BLUE_STAINED_GLASS = createBlocks(0.125F);
-    public static final CBBlock[] BLUE_STAINED_GLASS = createBlocks(0.125F);
-    public static final CBBlock[] PURPLE_STAINED_GLASS = createBlocks(0.125F);
-    public static final CBBlock[] MAGENTA_STAINED_GLASS = createBlocks(0.125F);
-    public static final CBBlock[] PINK_STAINED_GLASS = createBlocks(0.125F);
-    public static final CBBlock[] ICE = createBlocks(0.75F);
-    public static final CBBlock[] PACKED_ICE = createBlocks(0.75F);
-    public static final CBBlock[] BLUE_ICE = createBlocks(0.75F);
+    public static final CBBlock[] GLASS = createGlassBlocks(0.125F);
+    public static final CBBlock[] TINTED_GLASS = createGlassBlocks(0.25F);
+    public static final CBBlock[] WHITE_STAINED_GLASS = createGlassBlocks(0.125F);
+    public static final CBBlock[] LIGHT_GRAY_STAINED_GLASS = createGlassBlocks(0.125F);
+    public static final CBBlock[] GRAY_STAINED_GLASS = createGlassBlocks(0.125F);
+    public static final CBBlock[] BLACK_STAINED_GLASS = createGlassBlocks(0.125F);
+    public static final CBBlock[] BROWN_STAINED_GLASS = createGlassBlocks(0.125F);
+    public static final CBBlock[] RED_STAINED_GLASS = createGlassBlocks(0.125F);
+    public static final CBBlock[] ORANGE_STAINED_GLASS = createGlassBlocks(0.125F);
+    public static final CBBlock[] YELLOW_STAINED_GLASS = createGlassBlocks(0.125F);
+    public static final CBBlock[] LIME_STAINED_GLASS = createGlassBlocks(0.125F);
+    public static final CBBlock[] GREEN_STAINED_GLASS = createGlassBlocks(0.125F);
+    public static final CBBlock[] CYAN_STAINED_GLASS = createGlassBlocks(0.125F);
+    public static final CBBlock[] LIGHT_BLUE_STAINED_GLASS = createGlassBlocks(0.125F);
+    public static final CBBlock[] BLUE_STAINED_GLASS = createGlassBlocks(0.125F);
+    public static final CBBlock[] PURPLE_STAINED_GLASS = createGlassBlocks(0.125F);
+    public static final CBBlock[] MAGENTA_STAINED_GLASS = createGlassBlocks(0.125F);
+    public static final CBBlock[] PINK_STAINED_GLASS = createGlassBlocks(0.125F);
+    public static final CBBlock[] ICE = createBlocks(0.75F, true);
+    public static final CBBlock[] PACKED_ICE = createBlocks(0.75F, true);
+    public static final CBBlock[] BLUE_ICE = createBlocks(0.75F, true);
     public static final CBBlock[] SNOW_BLOCK = createBlocks(0.5F);
     public static final CBBlock[] MOSS_BLOCK = createBlocks(0.5F);
     public static final CBBlock[] CALCITE = createBlocks(null);
@@ -274,12 +274,34 @@ public class BlockList {
     public static final CBBlock[] SEA_LANTERN = createBlocks(null);
 
     private static CBBlock[] createBlocks(Float amplifier) {
+        return createBlocks(amplifier, false);
+    }
+
+    private static CBBlock[] createBlocks(Float amplifier, boolean isLesser) {
+        int maxCompressionLevel = 10;
+        CBBlock[] result = new CBBlock[maxCompressionLevel];
+        for (int i = 0; i < maxCompressionLevel; i++) {
+            float baseHardness = HARDNESS[i];
+            float baseResistance = RESISTANCE[i];
+
+            float factor = (amplifier == null ? 1 : amplifier) * (isLesser ? 0.5f : 1.0f);
+
+            float blockHardness = baseHardness * factor;
+            float blockResistance = baseResistance * factor;
+
+            result[i] = new CBBlock(BlockBehaviour.Properties.of(Material.STONE).strength(blockHardness, blockResistance), i, isLesser);
+        }
+        return result;
+    }
+
+    private static CBBlock[] createGlassBlocks(Float amplifier) {
         int maxCompressionLevel = 10;
         CBBlock[] result = new CBBlock[maxCompressionLevel];
         for (int i = 0; i < maxCompressionLevel; i++) {
             float blockHardness = amplifier != null ? HARDNESS[i] * amplifier : HARDNESS[i];
             float blockResistance = amplifier != null ? RESISTANCE[i] * amplifier : RESISTANCE[i];
-            result[i] = new CBBlock(BlockBehaviour.Properties.of(Material.STONE).strength(blockHardness, blockResistance), i);
+
+            result[i] = new CBBlock(BlockBehaviour.Properties.of(Material.GLASS).noCollission().strength(blockHardness, blockResistance), i, false);
         }
         return result;
     }