+++ changelog.md
@@ -1,8 +1,12 @@
-# v1.6.0 [ 1.19.4 ]
+# v1.6.1 [ 1.19.4 ]
 
 ## General
 
-- changed version scheming for clarity
+- Added `Lesser Compressed Blocks`, addresses #43
+    - crafted using a 2x2 recipe instead of a 3x3 recipe
+    - for now this applies to all levels, but may be changed in the future (open an issue or post a comment if you want
+      to
+      discuss this)
 
 - Compatible with: `[ 1.19.3, 1.19.4]`
 - ***NOT*** compatible with `[ 1.19, 1.19.1, 1.19.2 ]`